Moved some code for Collection configuration to the AbstractConfigurationConverter.
This code tried to guess the implementation class for the compound type of a collection
(for example 'A list of FooBars').

Now the ObjectWithFieldsConverter also uses this code to determine the implementation.

You can now specify the type of a component's field to be an interface. The name of the field
(or rather, the configuration tag for that parameter/field) will be used to determine the implementation,
just as with Collections.