add applicationXmlFile arg for services aswell;
remove default expression; we want to fail when the user supplies
a file, and we don't want to fail when he didn't. If he didn't,
check the default location and if the application.xml exists, use it.
This change is needed because there's no way to tell whether a parameter
is set by the user or set by the default (the user _could_ provide the default
value explicitly).